game-rent-api/requirements-dev.txt
-r requirements.txt

# Testes
pytest==9.0.2
pytest-django==4.12.0
pytest-cov==7.0.0
factory-boy==3.3.3
faker==40.11.0
